Minor bug fix re: 'data' object
In tutorial, there is a line where the example data is loaded:

load("example.Rdata")

The example data is stored in the variable `dat`.

apply_classifier() accepts 'data' and 'classifier' arguments. In the example, apply_classifier sets data to the example data (`dat`). Line 52 of functions.R had a typo wherein it accepted `dat` and not `data`. This caused failure when using a different dataset stored as a different variable name. This change resolves the problem.